some fix, used int64_t

+++ tasks/levonychev_i_mult_matrix_vec/common/include/common.hpp
@@ -7,8 +7,8 @@
 
 namespace levonychev_i_mult_matrix_vec {
 
-using InType = std::tuple<std::vector<int>, int, int, std::vector<int>>;
-using OutType = std::vector<int>;
+using InType = std::tuple<std::vector<int64_t>, int, int, std::vector<int64_t>>;
+using OutType = std::vector<int64_t>;
 using TestType = std::tuple<int, std::string>;
 using BaseTask = ppc::task::Task<InType, OutType>;
